movespeed and position changed to floats

+++ game_playground/Player.h
@@ -5,14 +5,14 @@ class Player : public sf::Drawable
 public:
 
 	int health;
-	int movespeed;
-	sf::Vector2i pos;
+	float movespeed;
+	sf::Vector2f pos;
 
 	sf::Texture texture;
 	sf::Sprite sprite;
 
 	Player() = delete;
-	Player(sf::Vector2i, int movespeed, int health, std::string texturePath);
+	Player(sf::Vector2f, float movespeed, int health, std::string texturePath);
 	~Player();
 
 	virtual void draw(sf::RenderTarget &, sf::RenderStates) const ;
